requirements.txt
langchain
langchain_openai
langchain-community
requests
python-dotenv
playwright
browser-use
supabase